+++ biblioteca-4/comunes/comprobaciones-general.php
@@ -74,7 +74,7 @@ function imprimeAvisosGenerales()
 
     foreach ($argumentos as $origen) {
         if (isset($_SESSION["avisosGenerales"][$origen]) && count($_SESSION["avisosGenerales"][$origen]) > 0) {
-            $_SESSION["avisosGenerales"][$origen] = array_unique($_SESSION["avisosGenerales"][$origen]);
+            $_SESSION["avisosGenerales"][$origen] = array_unique($_SESSION["avisosGenerales"][$origen], SORT_REGULAR);
             foreach ($_SESSION["avisosGenerales"][$origen] as $aviso) {
                 print "    <p class=\"$aviso[claseAviso]\">$aviso[texto]</p>\n";
             }