+++ biblioteca-4/comunes/comprobaciones-particular.php
@@ -17,23 +17,8 @@ function comprobaciones($origen, $tabla, $campo, $valor)
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "No ha seleccionado ningún registro.", "claseAviso" => "aviso-error"];
             $valor                                  = "";
         } else {
-            $pdo      = conectaDb();
-            $consulta = "SELECT COUNT(*)
-                         FROM $db[$tabla]
-                         WHERE id=:id_recibido";
-            $result = $pdo->prepare($consulta);
-            $result->execute([":id_recibido" => $valor]);
-            if (!$result) {
-                $mensaje = "Error en la consulta.";
-            } elseif ($result->fetchColumn() == 0) {
-                $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Registro no encontrado.", "claseAviso" => "aviso-error"];
-            } else {
-                $campoOk = true;
-            }
-            $pdo = null;
+            $campoOk = true;
         }
-    } elseif ($campo == "id") {                         // Cualquier tabla
-        $campoOk = true;
     } elseif ($campo == "usuario") {                    // Tabla Usuarios
         if ($valor == "") {
             $mensaje = "Debe escribir un nombre de usuario.";
@@ -185,14 +170,33 @@ function compruebaAvisosGenerales()
         if (is_array($argumentos[0])) {
             if (count($argumentos[0]) == 0) {
                 $_SESSION["avisosGenerales"][$origen][] = ["texto" => "No ha seleccionado ningún registro.", "claseAviso" => "aviso-error"];
-                return true;
             }
         } else {
             if ($argumentos[0] == "") {
                 $_SESSION["avisosGenerales"][$origen][] = ["texto" => "No ha seleccionado ningún registro.", "claseAviso" => "aviso-error"];
-                return true;
             }
         }
+    }
+
+    if ($tipoComprobacion == "validaLogin") {
+        $pdo      = conectaDb();
+        $consulta = "SELECT *
+                     FROM $argumentos[0]
+                     WHERE usuario=:usuario
+                    AND password=:password";
+        $result = $pdo->prepare($consulta);
+        $result->execute([":usuario" => recoge($argumentos[1]), ":password" => encripta(recoge($argumentos[2]))]);
+        if (!$result) {
+            $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
+        } else {
+            $valor = $result->fetch(PDO::FETCH_ASSOC);
+            if (!is_array($valor)) {
+                $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Nombre de usuario y/o contraseña incorrectos.", "claseAviso" => "aviso-error"];
+            } else {
+                $_SESSION["conectado"] = $valor["nivel"];
+            }
+        }
+        $pdo = null;
     }
 
     if ($tipoComprobacion == "yaExisteRegistro") {
@@ -212,34 +216,72 @@ function compruebaAvisosGenerales()
         $result->execute($parametros);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() > 0) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "El registro ya existe.", "claseAviso" => "aviso-error"];
-            return true;
         }
         $pdo = null;
     }
 
     if ($tipoComprobacion == "incluyeUsuarioRoot") {
         $pdo = conectaDb();
+        $consulta = "SELECT *
+                     FROM $db[usuarios]
+                     WHERE usuario='$cfg[rootName]'";
+        $result = $pdo->query($consulta);
+        if (!$result) {
+            $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
+        }
+        $valor = $result->fetch(PDO::FETCH_ASSOC);
         if (!is_array($argumentos[0])) {
             $argumentos[0] = [$argumentos[0] => "on"];
         }
-        foreach ($argumentos[0] as $indice => $valor) {
-            $consulta = "SELECT *
-                         FROM $db[usuarios]
+        foreach ($argumentos[0] as $indice => $valor2) {
+            if ($valor["id"] == $indice) {
+                $_SESSION["avisosGenerales"][$origen][] = ["texto" => "El usuario root no se puede borrar.", "claseAviso" => "aviso-error"];
+            }
+        }
+        $pdo = null;
+    }
+
+    if ($tipoComprobacion == "registrosExisten") {
+        $pdo = conectaDb();
+        if (!is_array($argumentos[1])) {
+            $consulta = "SELECT COUNT(*)
+                         FROM $argumentos[0]
                          WHERE id=:id";
             $result = $pdo->prepare($consulta);
-            $result->execute([":id" => $indice]);
+            $result->execute([":id" => $argumentos[1]]);
             if (!$result) {
                 $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-                return true;
+            } elseif ($result->fetchColumn() == 0) {
+                $_SESSION["avisosGenerales"][$origen][] = ["texto" => "El registro indicado no existe.", "claseAviso" => "aviso-error"];
             }
-            $valor = $result->fetch(PDO::FETCH_ASSOC);
-            if ($valor["usuario"] == $cfg["rootName"]) {
-                $_SESSION["avisosGenerales"][$origen][] = ["texto" => "El usuario root no se puede borrar o modificar.", "claseAviso" => "aviso-error"];
-                return true;
+        } else {
+            if (count($argumentos[1]) == 1) {
+                $consulta = "SELECT COUNT(*)
+                             FROM $argumentos[0]
+                             WHERE id=:id";
+                $result = $pdo->prepare($consulta);
+                $result->execute([":id" => key($argumentos[1])]);
+                if (!$result) {
+                    $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
+                } elseif ($result->fetchColumn() == 0) {
+                    $_SESSION["avisosGenerales"][$origen][] = ["texto" => "El registro indicado no existe.", "claseAviso" => "aviso-error"];
+                }
+            } else {
+                foreach ($argumentos[1] as $indice => $valor2) {
+                    $consulta = "SELECT COUNT(*)
+                             FROM $argumentos[0]
+                             WHERE id=:id";
+                    $result = $pdo->prepare($consulta);
+                    $result->execute([":id" => $indice]);
+                    if (!$result) {
+                        $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
+                    } elseif ($result->fetchColumn() == 0) {
+                        $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Al menos uno de los registros indicados no existe.", "claseAviso" => "aviso-error"];
+                    }
+                }
             }
         }
         $pdo = null;
@@ -264,11 +306,9 @@ function compruebaAvisosGenerales()
         $result->execute($parametros);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() > 0) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Ya existe un registro con esos mismos valores. Se muestran los datos originales (sin modificar).", "claseAviso" => "aviso-error"];
-            return true;
         }
         $pdo = null;
     }
@@ -290,11 +330,9 @@ function compruebaAvisosGenerales()
         $result->execute($parametros);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() == 0) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "No se han encontrado registros.", "claseAviso" => "aviso-error"];
-            return true;
         }
         $pdo = null;
     }
@@ -308,7 +346,6 @@ function compruebaAvisosGenerales()
         }
         if ($todosVacios) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Hay que rellenar al menos uno de los campos. No se ha guardado el registro.", "claseAviso" => "aviso-error"];
-            return true;
         }
     }
 
@@ -327,7 +364,6 @@ function compruebaAvisosGenerales()
         }
         if ($todosVacios) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Hay que rellenar al menos uno de los campos. Se muestran los datos originales (sin modificar).", "claseAviso" => "aviso-error"];
-            return true;
         }
     }
 
@@ -340,7 +376,6 @@ function compruebaAvisosGenerales()
         }
         if ($algunoVacio) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Hay que rellenar todos los campos. No se ha guardado el registro.", "claseAviso" => "aviso-error"];
-            return true;
         }
     }
 
@@ -350,19 +385,16 @@ function compruebaAvisosGenerales()
         $tabla = $argumentos[0];
         if (!in_array($tabla, $db["tablas"])) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "La tabla $tabla no existe.", "claseAviso" => "aviso-error"];
-            return true;
         }
         $consulta = "SELECT COUNT(*)
                      FROM $db[$tabla]";
         $result = $pdo->query($consulta);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() == 0) {
             $_SESSION["avisosGenerales"][$origen][]          = ["texto" => "No se ha creado todavía ningún registro.", "claseAviso" => "aviso-info"];
             $_SESSION["avisosGenerales"]["ocultaFormulario"] = true;
-            return true;
         }
         $pdo = null;
     }
@@ -373,7 +405,6 @@ function compruebaAvisosGenerales()
         $tabla = $argumentos[0];
         if (!in_array($tabla, $db["tablas"])) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "La tabla $tabla no existe", "claseAviso" => "aviso-error"];
-            return true;
         }
         $consulta = "SELECT COUNT(*)
                      FROM $db[prestamos]
@@ -381,12 +412,10 @@ function compruebaAvisosGenerales()
         $result = $pdo->query($consulta);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() == 0) {
             $_SESSION["avisosGenerales"][$origen][]          = ["texto" => "No hay préstamos pendientes de devolución.", "claseAviso" => "aviso-info"];
             $_SESSION["avisosGenerales"]["ocultaFormulario"] = true;
-            return true;
         }
         $pdo = null;
     }
@@ -397,19 +426,16 @@ function compruebaAvisosGenerales()
         $tabla = $argumentos[0];
         if (!in_array($tabla, $db["tablas"])) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "La tabla $tabla no existe", "claseAviso" => "aviso-error"];
-            return true;
         }
         $consulta = "SELECT COUNT(*)
                      FROM $db[$tabla]";
         $result = $pdo->query($consulta);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         if ($result->fetchColumn() >= $cfg["maxRegTabla"][$tabla]) {
             $_SESSION["avisosGenerales"][$origen][]          = ["texto" => "Se ha alcanzado el número máximo de registros que se pueden guardar. Por favor, borre algún registro antes.", "claseAviso" => "aviso-info"];
             $_SESSION["avisosGenerales"]["ocultaFormulario"] = true;
-            return true;
         }
         $pdo = null;
     }
@@ -422,7 +448,6 @@ function compruebaAvisosGenerales()
             || !checkdate(substr($despues, 5, 2), substr($despues, 8, 2), substr($despues, 0, 4))
             || $antes > $despues) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "La fecha de devolución <strong>$despues</strong> es anterior a la de préstamo <strong>$antes</strong>.", "claseAviso" => "aviso-error"];
-            return true;
         }
     }
 
@@ -437,15 +462,11 @@ function compruebaAvisosGenerales()
         $result->execute([":id" => $id_prestamo]);
         if (!$result) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "Error en la consulta.", "claseAviso" => "aviso-error"];
-            return true;
         }
         $prestado = $result->fetchColumn();
         if ($prestado > $devuelto) {
             $_SESSION["avisosGenerales"][$origen][] = ["texto" => "La fecha de devolución <strong>$devuelto</strong> es anterior a la de préstamo <strong>$prestado</strong>.", "claseAviso" => "aviso-error"];
-            return true;
         }
         $pdo = null;
     }
-
-    return false;
 }