+++ biblioteca-4/tablas/db-prestamos/borrar-2.php
@@ -13,6 +13,8 @@ borraAvisosExcepto();
 
 [$id] = compruebaAvisosIndividuales("borrar-2", "prestamos", "id[]");
 
+compruebaAvisosGenerales("borrar-2", "registrosExisten", "prestamos", $id);
+
 if (hayErrores("borrar-2")) {
     header("Location:borrar-1.php");
     exit();